requirements.txt
pyyaml>=6.0
tqdm>=4.66
soundfile>=0.12
google-generativeai>=0.3.0
funasr>=1.0.0
onnxruntime>=1.16.0
fastapi>=0.110
uvicorn>=0.23
python-multipart>=0.0.9
jinja2>=3.1
requests>=2.31
python-dotenv>=1.0.1
